Update specs to use correct stack instantiation.
